moved client builder into its own module
